Add check for 0 length table for function '_parseSpectralResolution' and '_parsePolarizations'

+++ ALMAxmatch/core.py
@@ -486,28 +486,29 @@ class archiveSearch:
         """
         for tar in self.targets:
             table = self.queryResults[tar]
-            if type(table['Frequency resolution'][0]) != np.float64:
-                msg = 'Dev alert: "Frequency resolution" may have more than '
-                msg += 'one entry per observation so it may not be wise to '
-                msg += 'completely replace it in _parseSpectralResolution '
-                msg += 'anymore.'
-                print(msg)
-            targetRes = list()
-            for i in range(len(table)):
-                freqStr = table['Frequency support'][i]
-                freqStr = freqStr.split('U')
-                rowRes = list()
-                for j in range(len(freqStr)):
-                    resolution = freqStr[j].split(',')
-                    resolution = u.Quantity(resolution[1])
-                    resolution = resolution.to('kHz')
-                    rowRes.append(resolution.value)
-                targetRes.append(rowRes)
+            if len(table) > 0:
+                if type(table['Frequency resolution'][0]) != np.float64:
+                    msg = 'Dev alert: "Frequency resolution" may have more than '
+                    msg += 'one entry per observation so it may not be wise to '
+                    msg += 'completely replace it in _parseSpectralResolution '
+                    msg += 'anymore.'
+                    print(msg)
+                targetRes = list()
+                for i in range(len(table)):
+                    freqStr = table['Frequency support'][i]
+                    freqStr = freqStr.split('U')
+                    rowRes = list()
+                    for j in range(len(freqStr)):
+                        resolution = freqStr[j].split(',')
+                        resolution = u.Quantity(resolution[1])
+                        resolution = resolution.to('kHz')
+                        rowRes.append(resolution.value)
+                    targetRes.append(rowRes)
 
-            table.remove_column('Frequency resolution')
+                table.remove_column('Frequency resolution')
 
-            table['Frequency resolution'] = targetRes
-            table['Frequency resolution'].unit = 'kHz'
+                table['Frequency resolution'] = targetRes
+                table['Frequency resolution'].unit = 'kHz'
 
     def _parseLineSensitivities(self):
         """Parses all line sensitivity information into a more useful form.
@@ -576,26 +577,27 @@ class archiveSearch:
         """
         for tar in self.targets:
             table = self.queryResults[tar]
-            if type(table['Pol products'][0]) != str:
-                msg = 'Dev alert: "Pol products" column may have more than '
-                msg += 'one entry per observation so it may not be wise to '
-                msg += 'completely replace it in _parseSpectralResolution '
-                msg += 'anymore.'
-                print(msg)
-            targetPol = list()
-            for i in range(len(table)):
-                freqStr = table['Frequency support'][i]
-                freqStr = freqStr.split('U')
-                rowPol = list()
-                for j in range(len(freqStr)):
-                   polarization  = freqStr[j].split(',')
-                   polarization = polarization[4].strip(' ]')
-                   rowPol.append(polarization)
-                targetPol.append(rowPol)
+            if len(table) > 0:
+                if type(table['Pol products'][0]) != str:
+                    msg = 'Dev alert: "Pol products" column may have more than '
+                    msg += 'one entry per observation so it may not be wise to '
+                    msg += 'completely replace it in _parseSpectralResolution '
+                    msg += 'anymore.'
+                    print(msg)
+                targetPol = list()
+                for i in range(len(table)):
+                    freqStr = table['Frequency support'][i]
+                    freqStr = freqStr.split('U')
+                    rowPol = list()
+                    for j in range(len(freqStr)):
+                       polarization  = freqStr[j].split(',')
+                       polarization = polarization[4].strip(' ]')
+                       rowPol.append(polarization)
+                    targetPol.append(rowPol)
 
-            table.remove_column('Pol products')
+                table.remove_column('Pol products')
 
-            table['Pol products'] = targetPol
+                table['Pol products'] = targetPol
 
     def dumpSearchResults(self, target_data, bands,
                           unique_public_circle_parameters=False,
